investment entity Long변수들 정확성을 위해 BigDecimal로 타입 변경

+++ src/main/java/com/example/seoultechInvestment/DTO/EndedInvDTOfromDB.java
@@ -10,6 +10,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -25,9 +26,9 @@ public class EndedInvDTOfromDB {
     @Pattern(regexp = "[0-9]*D$")
     private String holdTerm;
     @NotNull()
-    private double sellPrice;
+    private BigDecimal sellPrice;
     @NotNull
-    private double earningRate;
+    private BigDecimal earningRate;
     @NotNull
     @StatusPattern
     private ProgressStatus status;
